+++ ListableUI/Sources/Appearance.swift
@@ -22,7 +22,7 @@ public struct Appearance : Equatable
         
     /// Creates a new appearance object with the provided options.
     public init(
-        backgroundColor : UIColor = .white,
+        backgroundColor : UIColor = Self.defaultBackgroundColor,
         showsScrollIndicators : Bool = true,
         configure : (inout Self) -> () = { _ in }
     ) {
@@ -31,5 +31,23 @@ public struct Appearance : Equatable
         self.showsScrollIndicators = showsScrollIndicators
         
         configure(&self)
+    }
+    
+    /// The default background color for the `Appearance`.
+    public static var defaultBackgroundColor : UIColor {
+        if #available(iOS 13.0, *) {
+            return UIColor { traits in
+                switch traits.userInterfaceStyle {
+                case .unspecified, .light:
+                    return .white
+                case .dark:
+                    return .black
+                @unknown default:
+                    return .white
+                }
+            }
+        } else {
+            return .white
+        }
     }
 }